backend/requirements.txt
uvicorn
fastapi
pillow-heif
supabase
python-multipart
Pillow
